Server configuration json and pickup/drop control
Servers can save and load a custom json file.  In this example we're setting a value for the max number of droppable items.   The backend now has control over individual players ability to pick up and drop items on each server.  This is most interesting for group instances, which can assign roles to users with these permissions.